+++ castle/cms/static/components/manage-archives.js
@@ -33,24 +33,34 @@ require([
         items: [],
         path: InitialPath || '/',
         selected: [],
-        previousSelected: null
+        previousSelected: null,
+        page: 1
       };
     },
 
     componentDidMount: function(){
-      this.load();
+      this.load(true);
     },
 
-    load: function(){
+    load: function(checkpage){
       var that = this;
       utils.loading.show();
+      var thepage = that.state.page;
+      if(checkpage) {
+        var qspage = getQueryVariable('page');
+        if(qspage !== undefined) {
+          that.setState({page: qspage});
+          thepage = qspage;
+        }
+      }
       var url = window.location.origin + window.location.pathname;
       $.ajax({
         url: url,
         data: {
           api: 'true',
           method: 'list',
-          path: that.state.path
+          path: that.state.path,
+          page: thepage
         },
         cache: false
       }).done(function(items){
@@ -89,12 +99,33 @@ require([
           }
         }).done(function(){
           utils.loading.hide();
-          that.load();
+          that.load(true);
         }).fail(function(){
           utils.loading.hide();
           alert('error deleting');
         });
       }
+    },
+
+    navigatePrevClicked: function(item, e) {
+      var curpage = this.state.page;
+      var newpage = curpage - 1;
+      if(newpage < 1) {
+        newpage = 1;
+      }
+      var that = this;
+      this.setState({page: newpage}, function(){
+	 that.load(false);
+      });
+    },
+
+    navigateNextClicked: function(item, e) {
+      var curpage = this.state.page;
+      var newpage = curpage + 1;
+      var that = this;
+      this.setState({page: newpage}, function(){
+        this.load(false)
+      });
     },
 
     renderItem: function(item, idx){
@@ -219,8 +250,7 @@ require([
         });
       }
       return D.div({}, [
-        parts,
-        D.span({ className: 'pull-right'}, that.state.items.length + ' items in folder')
+        parts
       ]);
     },
 
@@ -300,12 +330,29 @@ require([
       ]);
     },
 
+    renderNav: function(){
+      var that = this;
+      return D.div({}, [
+        D.button({
+          className: 'plone-btn plone-btn-default',
+          disabled: that.state.page <= 1,
+          onClick: that.navigatePrevClicked
+        }, 'Previous Page'),
+        D.button({
+          className: 'plone-btn plone-btn-default pull-right',
+          disabled: that.state.items.length < 100,
+          onClick: that.navigateNextClicked
+        }, 'Next Page')
+      ])
+    },
+
     render: function(){
       var that = this;
       return D.div({}, [
         that.renderButtons(),
         that.renderTable(),
-        that.renderSelected()
+        that.renderSelected(),
+        that.renderNav()
       ]);
     }
   });